+++ src/AhkSymbol/TAhkSymbolIn.ts
@@ -1,10 +1,8 @@
 import type { CAhkClass, CAhkClassGetSet, CAhkClassInstanceVar } from './CAhkClass';
 import type { CAhkFunc } from './CAhkFunc';
+import type { CAhkHotKeys } from './CAhkHotKeys';
 import type { CAhkHotString } from './CAhkHotString';
-import type {
-    CAhkComment,
-    CAhkHotKeys,
-} from './CAhkLine';
+import type { CAhkComment } from './CAhkLine';
 import type { CAhkCase, CAhkDefault, CAhkSwitch } from './CAhkSwitch';
 import type { TLineClass } from './TLineClass';