+++ src/tools/Built-in/9_HotStrings_Options/HotStringsOptions.data.ts
@@ -32,9 +32,6 @@ export type THotStringsOptionsList = Readonly<{
     parseFn: (s: string) => THotStringsOptions | null,
 }>;
 
-/**
- * after initialization clear
- */
 export const HotStringsOptionsList: THotStringsOptionsList[] = [
     {
         humanName: '* (asterisk)',